refactor: consolidate duplicate _protocols.py into core/protocols.py
Remove top-level _protocols.py (AsyncEmbeddingModel, EmbeddingModel,
LLMBackend) — these were duplicated in core/protocols.py. Update the
two importers (cache.py, memory_analyzer.py) to use core/protocols.
