+++ src/mnemotree/analysis/memory_analyzer.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 import asyncio
 from typing import Any
 
-from .._protocols import AsyncEmbeddingModel as Embeddings
 from ..core.models import MemoryType
+from ..core.protocols import EmbeddingModel as Embeddings
 from ..utils.serialization import json_dumps_safe
 from .analyzers import ConceptAnalyzer, EmotionAnalyzer, MemoryClassifierAnalyzer, PatternAnalyzer
 from .models import (
